requirements.txt
Django
django-ckeditor
django-cleanup
django-crspy-forms
django-js-asset